Add beast and basics sources to lib, and tests
This may be a more controvercial change, since we have to balance scope
creep, testing, and library completeness.
* The intention is to remove these files from rippled. Any file removed
there MUST be a part of the library. I've chosen to err on the side of
including more in the library. If we go the other way, we'll need to be
much more granular in deciding which files are moved and which are kept.
Any files kept in rippled should be deleted from ripple-libpp.
* This change is almost certainly not exhaustive, since our test
coverage is not exhaustive.